// src/openapi/models/s3-pagination-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { DocumentDto } from './document-dto';

/**
 * 
 * @export
 * @interface S3PaginationDto
 */
export interface S3PaginationDto {
    /**
     * 
     * @type {Array<DocumentDto>}
     * @memberof S3PaginationDto
     */
    documents: Array<DocumentDto>;
    /**
     * 
     * @type {string}
     * @memberof S3PaginationDto
     */
    currentContinuationToken?: string;
    /**
     * 
     * @type {string}
     * @memberof S3PaginationDto
     */
    nextContinuationToken?: string;
    /**
     * The size of the page
     * @type {number}
     * @memberof S3PaginationDto
     */
    size: number;
    /**
     * The size of the returned elements of this page
     * @type {number}
     * @memberof S3PaginationDto
     */
    totalElements: number;
}